requirements.txt
Django==4.2.13
paho-mqtt
celery
gunicorn
channels
channels_redis
djangorestframework
markdown==3.4.3
django-filter==23.2
python-decouple
unipath
django-multiselectfield
daphne
django-import-export
whitenoise
paho-mqtt
apscheduler
requests
python-decouple
drf-yasg[validation]
paho-mqtt

